syncs up test files
